Converted ImageStore to a singleton and squashed its public API
