+++ Sources/UI/Views/TextCell.swift
@@ -22,8 +22,8 @@ open class TextCell: BlockCell {
         return textView
     }
     
-    override open func configure(with block: Block, imageStore: ImageStore) {
-        super.configure(with: block, imageStore: imageStore)
+    override open func configure(with block: Block) {
+        super.configure(with: block)
         
         guard let textBlock = block as? TextBlock else {
             textView.isHidden = true
